requirements.txt
faiss-cpu
numpy
tqdm
requests
sentence-transformers
scikit-learn
